+++ jablib/src/main/java/org/jabref/logic/git/GitHandler.java
@@ -204,6 +204,7 @@ public class GitHandler {
                    .setAmend(amend)
                    .setAllowEmpty(false)
                    .setMessage(commitMessage)
+                   .setSign(false)
                    .call();
             }
         }